requirements.txt
beautifulsoup4==4.12.3
Django==3.2.10
django-allauth==0.58.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==9.5.0
httpx==0.27.0
huggingface-hub==0.24.0
llama-index==0.10.56
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-readers-web==0.1.23
llama-parse==0.4.9
lockfile==0.12.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.2.1
requests==2.31.0
virtualenv==20.25.1
psycopg2==2.9.9
faiss-cpu==1.8.0
groq==0.9.0
sentence-transformers
setuptools==71.0.4
gunicorn==22.0.0
pyPDF2==3.0.1
docx==0.2.4
python-docx==1.1.2